authorization filter

+++ src/Controller/DevicesController.php
@@ -217,8 +217,8 @@ public function ajaxdata() {
                 $this->Flash->error(__('The device could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Devices->Customers->find('list', ['limit' => 200]);
-        $providers = $this->Devices->Providers->find('list', ['limit' => 200]);
+        $customers = $this->Devices->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
+        $providers = $this->Devices->Providers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('device', 'customers', 'providers'));
         $this->set('_serialize', ['device']);
     }
@@ -235,6 +235,12 @@ public function ajaxdata() {
         $device = $this->Devices->get($id, [
             'contain' => []
         ]);
+		if($device['customer_id']!= $this->loggedinuser['customer_id'])
+		{
+			 $this->Flash->success(__('You are not Authorized.'));
+			 return $this->redirect(['action' => 'index']);
+			
+		}
         if ($this->request->is(['patch', 'post', 'put'])) {
             $device = $this->Devices->patchEntity($device, $this->request->data);
             if ($this->Devices->save($device)) {
@@ -245,8 +251,8 @@ public function ajaxdata() {
                 $this->Flash->error(__('The device could not be saved. Please, try again.'));
             }
         }
-        $customers = $this->Devices->Customers->find('list', ['limit' => 200]);
-        $providers = $this->Devices->Providers->find('list', ['limit' => 200]);
+        $customers = $this->Devices->Customers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
+        $providers = $this->Devices->Providers->find('list', ['limit' => 200])->where(['customer_id' => $this->loggedinuser['customer_id']])->orwhere(['customer_id' => '0']) ;
         $this->set(compact('device', 'customers', 'providers'));
         $this->set('_serialize', ['device']);
     }
@@ -262,12 +268,20 @@ public function ajaxdata() {
     {
         $this->request->allowMethod(['post', 'delete']);
         $device = $this->Devices->get($id);
-        if ($this->Devices->delete($device)) {
-            $this->Flash->success(__('The device has been deleted.'));
-        } else {
-            $this->Flash->error(__('The device could not be deleted. Please, try again.'));
-        }
-
+		
+		if($device['customer_id'] = $this->loggedinuser['customer_id'])
+		{
+	        if ($this->Devices->delete($device)) {
+	            $this->Flash->success(__('The device has been deleted.'));
+	        } else {
+	            $this->Flash->error(__('The device could not be deleted. Please, try again.'));
+	        }
+		}
+		else
+		{
+		   	    $this->Flash->error(__('You are not authorized'));
+			
+		}
         return $this->redirect(['action' => 'index']);
     }
